+++ Ellipse.lua
@@ -11,9 +11,12 @@ local TWO_PI = math.pi * 2
 --- @field protected _radiusY number Vertical radius
 local Ellipse = utils.class("loveshape.Ellipse", Shape)
 
+--- Create a new ellipse with the specified dimensions.
+---
+--- If `pointsCount` is not specified, a default value is computed automatically.
 --- @param radiusX number Horizontal radius
 --- @param radiusY number Vertical radius
---- @param pointsCount integer Number of points used to draw the ellipse
+--- @param pointsCount? integer Number of points used to draw the ellipse
 --- @return loveshape.Ellipse
 --- @nodiscard
 function Ellipse.new(radiusX, radiusY, pointsCount)
@@ -25,11 +28,15 @@ end
 
 --- @param radiusX number
 --- @param radiusY number
---- @param pointsCount integer
+--- @param pointsCount? integer
 --- @protected
 function Ellipse:_init(radiusX, radiusY, pointsCount)
   assert(type(radiusX) == "number" and radiusX > 0)
   assert(type(radiusY) == "number" and radiusY > 0)
+
+  if not pointsCount then
+    pointsCount = utils.computeEllipsePoints(radiusX, radiusY)
+  end
 
   Shape._init(self, pointsCount)
   self:setRadius(radiusX, radiusY)